Consider a genex as a single command arg, even if it contains a list.
Up to now, cmake code such as

 add_custom_command( ... $<1:FOO;BAR> ...)

would split the arguments into

 ... $<1:FOO BAR> ...

That is, the generator expression was broken up, and neither part is a
generator expression on its own anymore. As the determination of whether
to add the 'current source directory' as a prefix for includes
in target_include_directories depends on whether the argument is a
generator expression, this meant that code such as

 target_include_directories(foo PRIVATE $<1:/bar/foo;/bat/mar/>)

was not handled correctly.