Add EACH and ALL generator expressions.
The EACH expression allows inline transformation of list elements
while evaluating generator expressions.

 $<EACH:one;two,TOUPPER>

if there was such a $<TOUPPER:input> expression, would result in
the list:

 ONE;TWO

The expression

 $<EACH:ON;OFF,BOOL>

results in the list

 1;0

Where a generator expression requires more than one argument, they
can be supplied:

 $<EACH:one;two,STREQUAL,one>

will result in

 1;0

because $<STREQUAL:one,one> and $<STREQUAL:one,two> will be invoked.

There is currently no way to specify which position the elements from
the list should be inserted into the parameters of the invoked
expression. Currently it is always appended. Some special marker
like $<_1>, inspired by std::bind might be appropriate.

The ALL expression sends all elements of a list to a specified
generator expression:

 $<ALL:AND,0;0;0>

invokes $<AND:0,0,0> for example.

The real use for these is things like normalizing and escaping paths
so that we might have:

 $<EACH:$<TARGET_PROPERTY:foo,INCLUDE_DIRECTORIES>,CONVERT_TO_UNIX_SLASHES>

to get the includes from foo and transform them appropriately.