Allow the SYSTEM keyword in target_link_libraries.
As target_link_libraries now also processes include directories
from the interface of targets on the right hand side, it makes
sense to allow to specify that the includes should be treated
as system includes.

This reduced two lines to one. Before:

 target_link_libraries(foo PRIVATE Qt5::Core)
 target_include_directories(foo SYSTEM PRIVATE $<TARGET_PROPERTY:Qt5::Core,INTERFACE_INCLUDE_DIRECTORIES>)

After:

 target_link_libraries(foo SYSTEM PRIVATE Qt5::Core)

It could even make sense to treat all IMPORTED targets on the right
hand side as a source of system includes, as they are external
dependencies.