Genex: Add API to evaluate excluded content.
A generator expression may conditionally generate content by evaluating
boolean expressions in its left-hand-side, and optionally also on its
right-hand-side via nesting of further conditions.  The boolean
conditions are evaluated to either of the output expressions '0' or '1'
depending on the context.

 $<$<COND>:content>
 $<$<AND:$<COND1>,$<COND2>>:content>
 $<$<COND1>:$<$<COND2>:content>>

An interesting property of generator expressions is that there is
exactly one position within them which can contain 'output content',
even if there are many complex conditions, and even if there are
permutations of possible output content.  That is, there is no IF-ELSE
construct in generator expressions (currently.  Nothing prevents it,
but it would be a departure from the expressive intent of the language
as designed), and even when multiple permutations may be generated, all
must be defined in one position in the generator expression, such as

 $<$<COND>:output_$<GENEX>_suffix>

If the condition for generating the content is evaluated to '0', no
content is output.  However, in some scenarios, we would like to know
the content that would have been generated if the condition had
instead been true.

This commit enables evaluation of the content that is excluded by a
generator expression in cases where there are exactly two possible
outputs, and one of those outputs is emptiness.  That is, the position
of the generator expression which contains the output must be 'plain
text' and must not contain any generator expressions itself.

Add new EvaluateExcluded() methods on cmCompiledGeneratorExpression to
return the 'content' of the expression where possible.  It is possible
to determine the excluded content in cases where the 'top level' of
the expression is a boolean conditional output expression (ie, a '0'
or a '1'), and the content is plain text.  If the left-hand-side of
the top-level expression evaluates to a '0', and each nested level of
the right-hand-side is a single boolean generator expression then
traverse the tree of conditions until plain text content is reached,
transforming '0' expressions into '1' at the relevant level as
needed.

At the termination of the traversal, if we needed to transform any
'0' into a '1', then the content was excluded and should be returned
from this method.  If no '0' was encountered, then the content was
not excluded in a normal evaluation and empty content is returned.
'0' and '1' expressions which are not at the appropriate level do
not participate in this end-output determination.  For example, the
'0' within the 'OR' expression here is not relevant:

 $<1:$<OR:$<BOOL:$<0:>>,$<1:1>>:content>

The 'OR' generates a '1' in 'normal' evaluation and so content is
generated by the entire expression, and so in 'excluded' evaluation,
which is 'the opposite', the entire expression generates an empty
result.

This algorithm assumes that the boolean conditions are
satisfiable.  That is, a generator expression like

 $<$<AND:$<CONFIG:Debug>,$<NOT:$<CONFIG:Debug>>>:content>

will not in reality produce content for any configuration, but the
content will still be reported as excluded by the generator
expression.

Generator expressions which have 'simple' right-side nesting could
instead be 'normalized' into AND expressions, but the implementation
of this algorithm is so simple that I didn't even look into doing
that.