cmTarget: Add API to retrieve sources excluded by genexes.
This allows determining what files were excluded from compilation
into a target where the conditions are specified with sufficiently
simple generator expressions.

For example in the following list, we can determine on a Linux
platform with a Release build that 'foo_win.cpp', 'foo_debug.cpp' and
'foo_win_debug.cpp' are "part of the project", but not part of this
build, as is desireable for IDE integation:

 add_executable(myexe
   main.cpp
   $<$<PLATFORM_ID:Windows>:foo_win.cpp>
   $<$<CONFIG:Debug>:foo_debug.cpp>
   $<$<AND:$<PLATFORM_ID:Windows>,$<CONFIG:Debug>>:foo_win_debug.cpp>
 )

Excluded files are listed only if they exist, because it is assumed
that files generated during the build for a particular platform or
configuration will not be generated if the platform or configuration
doesn't match.  Additionally, source files, not generated
files are likely more relevant to IDEs.